Sort out custom string types
Better reflect the intention of `llvm::Str` being just an alternative implementation of `std::ffi::CStr` that uses a slightly different definition and doesn't perform unnecessary length calculations but that both types store the same data (a raw pointer to a null terminated string). Because of this, I switched to using the generic `T: Borrow<Str>` instead of trait object`&AsRef<Str>` since Str is just another form of a borrowed CString.